requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
paramiko==2.12.0
sshtunnel==0.4.0
opencv-python-headless==4.8.1.78
Pillow==10.1.0
pytesseract==0.3.10
numpy==1.26.2
pytz==2023.3
